Add files via upload
Write an application for a bottle vending machine in which a customer inserts an amount and
selects the cold drink he/she wants. We assume that the vending machine has five following
beverages (7up, Fanta, Sprite, Pepsi, Coke). Show the beverages and their prices (of your choice)
to the user. Display the beverage name selected by the user, its price and the remaining amount
after purchase of a cold drink. Machine would display an error message: “Out of money” if the
amount entered by the user is less than the cost of beverage, he/she wants to purchase. Following
is just a sample, values in the red are user inputs.
====================
Welcome to the vending machine. We have following drinks.
Enter 1 for Pepsi, price = Rs. 80
Enter 2 for Coke, price = Rs. 80
Enter 3 for 7up, price = Rs. 75
Enter 4 for Sprite, price = Rs. 79
Enter 5 for Fanta, price = Rs. 76
Which drink do you want? 2
Enter the amount you have? 70
Out of money, you cannot buy a Coke