Add files via upload
A private electricity company needs an application that allows their customers to calculate their electricity bills. A customer will input units consumed and his/her bill will be calculated according
to the following criteria:
• For first 40 units Rs. 8/unit
• For next 60 units Rs. 10/unit
• For next 60 units Rs. 14/unit
• For unit above 160 Rs. 18/unit
Furthermore, fuel surcharges of 20% and Govt. Tax of 15% will also be included to a customer’s
bill. Calculate and display the total bill of a customer.